First attempt to move comprehensively from lists of fake information to dicts of real info.
The main reason to do that was that, since zone groups can change and
rearrange, it makes less sense to try to keep an ordered array of elements
which would constantly have to be managed so they don't become inconsistent
and instead just keep dictionaries.

I think that it may make sense to move this information into a nice struct
rather than maintain separate lists. But that is the next iteration.